requirements.txt
# Sika Sentinel — Python dependencies
# Install: pip install -r requirements.txt

# ── Demo UI ───────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Data validation ───────────────────────────────────────────────────────────
pydantic>=2.7.0

# ── Hedera Agent Kit — LangChain-compatible Hedera tooling ────────────────────
# Primary execution path: intent parsing, HBAR transfers, HCS messaging
hedera-agent-kit>=1.0.0

# ── LangChain — intent parsing orchestration ──────────────────────────────────
langchain>=0.2.0
langchain-openai>=0.1.0       # default LLM provider; swap for langchain-anthropic if preferred

# ── Hedera Python SDK — direct SDK access for HCS client & fallback ───────────
hedera-sdk-py>=2.0.0

# ── HTTP client — mirror node reads for audit replay ─────────────────────────
httpx>=0.27.0

# ── Environment variables ─────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
